swap kotlin serialization with moshi

+++ feature/search/src/main/kotlin/br/com/timetotrade/search/domain/model/SearchResult.kt
@@ -7,5 +7,5 @@ data class SearchResult(
     val exchDisp: String,
     val industry: String,
     val exchange: String,
-    val score: Int,
+    val score: Double,
 )
